Cambios a eliminar trabajador

+++ LOGEliminar_Trabajador.php
@@ -9,38 +9,35 @@
 
 
     $Eliminar_id=$_GET['id'];
-    if($conn_sis){
-        $consulta_T ="SELECT COUNT(*) AS ter from Permisos WHERE id_empleado='$Eliminar_id'";
-        $trab1 =sqlsrv_query($conn_sis,$consulta_T);
-        $que = sqlsrv_fetch_array($trab1);
-        $qu=$que["ter"];
-
-        if($qu==0){
-            $borrar = "DELETE FROM Empleados WHERE id_empleado='$Eliminar_id'";
-            $query=sqlsrv_query($conn_sis,$borrar);
-            if($query){
-                echo "<script type='text/javascript'>alert('Trabajador eliminado');
-                        window.location.href='lista_trabajador.php';
-                    </script>";
+    $queryPermisos="SELECT * FROM Permisos where id_empleado='$Eliminar_id'";
+            $resPer= sqlsrv_query($conn_sis, $queryPermisos);
+            if( $resPer === false) {
+                    die( print_r( sqlsrv_errors(), true) );
             }
-            else{
-                
+            while( $rowPer = sqlsrv_fetch_array($resPer) ) {
+                $idPer=$rowPer["id_empleado"];
+                $Per1=$rowPer["permiso1"];
+            }
+    if($conn_sis && $Per1==0){
+            $del_permisos="DELETE FROM Permisos WHERE id_empleado='$Eliminar_id' and id_empleado!=1";
+            $ejecutar=sqlsrv_query($conn_sis,$del_permisos);        
+            
+            if($ejecutar){
+                $borrar = "DELETE FROM Empleados WHERE id_empleado='$Eliminar_id'";
+                $query=sqlsrv_query($conn_sis,$borrar);
+                echo '<script>alert("Trabajador eliminado");window.location.href="lista_trabajador.php";</script>';                
+            }
+            else{                
                 echo "<script type='text/javascript'>alert('No se pudo eliminar el trabajador');
                 window.location.href='lista_trabajador.php';
             </script>";
             }
+    }
+    else{
+        echo "<script type='text/javascript'>alert('No Puedes eliminar a un SuperAdministrador');
+        window.location.href='lista_trabajador.php';</script>";
+    }
 
-        }else{
-            echo "<script type='text/javascript'>confirm('No se puede eliminar el trabajador ya que esta reguistrado en los permisos ¿Deseas desavilitar al trabajador?');
-            window.location.href='lista_trabajador.php';
-        </script>";
-
-        }
-    }else{
-    echo "<script type='text/javascript'>alert('No se puede ecceder a la base de datos');
-            window.location.href='lista_sucursal.php';  
-            </script>";
-            }
     sqlsrv_close($conn_sis);
